+++ src/components/[guild]/activity/ActivityLogFiltersBar/components/FilterTag/index.ts
@@ -1,0 +1,3 @@
+import FilterTag from "./FilterTag"
+
+export default FilterTag
